Add IAR C/C++ Compiler support.
There is a lot of noise because of a couple warnings it can emit
warnings about C-style casts in some configurations (--remarks), and
it treats `(void) foo;` as such a cast, which I was using to mark
variables as unused.

+++ test/array-param.c
@@ -4,15 +4,15 @@
 #include <assert.h>
 
 static void
-clear_buffer (int* buffer_size, unsigned char buffer[HEDLEY_ARRAY_PARAM(*buffer_size)]) {
-  int p;
+clear_buffer (size_t* buffer_size, unsigned char buffer[HEDLEY_ARRAY_PARAM(*buffer_size)]) {
+  size_t p;
   for (p = 0 ; p < *buffer_size ; p++)
     buffer[p] = 0;
 }
 
 int main(int argc, char* argv[HEDLEY_ARRAY_PARAM(argc)]) {
   unsigned char foo[8192];
-  int foo_l = (int) sizeof(foo);
+  size_t foo_l = sizeof(foo);
 
   assert(argv[argc - 1] != NULL);
